Changed the name of the files for the tsx and css to pair them. Also I changed the name of the component to be like the filename. I also added some documentation info
