+++ src/components/UpcomingEvents.tsx
@@ -5,7 +5,7 @@
  * Just on the landing page as the user opens the website 
  */
 
-import Programs_Section_Card from './Progams_Section_card';
+import ProgramSectionCard from './ProgramSectionCard';
 import '../styles/UpcomingEvents.css'
 
 function UpcomingEvents() {
@@ -49,18 +49,18 @@ function UpcomingEvents() {
         <div className="divider2"></div>
 
         <div className='info-cards'>
-          <Programs_Section_Card
-            image = "../assets/teacher-kids-selfie.png"
+          <ProgramSectionCard
+            image = "vite.svg"
             info = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla efficitur viverra sollicitudin."
             link = "programs"
           />
-          <Programs_Section_Card
+          <ProgramSectionCard
             image = "https://via.placeholder.com/300x180"
             info = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla efficitur viverra sollicitudin."
             link = "programs"
           />
-          <Programs_Section_Card
-            image = "https://via.placeholder.com/300x180"
+          <ProgramSectionCard
+            image = "../src/assets/kids-riding-hippo.png"
             info = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla efficitur viverra sollicitudin."
             link = "programs"
           />
